Uploaded shape files
Added shape file data of the polygons. If it is not usable, I can upload the outlines of each polygon.
In addition added two images for visibility of the different polygons.

I am not entirely sure, if it yields more advantages using the shapes instead of the sections of the polygons, what do you think?
Image analysis on shape instead/in addition to sections? To address, not only geometry, but also shape(form) of the tips etc.