Updated code
Emplaced ploygon segments under the big outline/trace figure for better evaluation of the dyke shape
I am aware, that is ugly the code as well as the colours, but maybe we can come up with something.

To make it work: run the code until the Large figure section, then run the live script in the FullData_shape and then go back to the dykeproof file and just run the large figure section. Essentially, I have not imported the polygone code, hence you need the variables in your workspace to make the code work. Apologies for the ugly way.